Fix mid-section alignment issues
The part displaying information with small images was not centered
which made the page look bad. The problem was that they were not
wrapping across which made the rightmost one be included in the
padding so it was not centered with the others. It was fixed by
allowing for wrapping.