Delete unnecessary css
When the previous bug was fixed, at first the whole mid section was
made a flex display type with columnal direction and align-items: center
in order to try to fix the centering issue. This fixed it but there
were still bugs in the padding. When those were fixed it also accounted
for the original centering issues, so the whole mid section no longer
needed to have a display type of flex - it was deleted.